🎨 Rewrite with custom hooks
